changing some 'is' to '=='
There are many parameter inputs. Presumably one would create a 'config' file, and if the config file has something like:
`cfg = {mcp : 'maxstat'}`

and when you use in code, it does not work because 'is' is not the same as '==' .
`cfg.get('mcp')` will show 'maxstat' but `mcp = cfg.get('mcp')` will throw an error since it is not comparing the string.